chore: complete structure overhaul
