+++ packages/ucd-store/src/index.ts
@@ -1,18 +1,16 @@
-export type { DownloadError, MirrorOptions, MirrorResult } from "./mirror";
-
-export { mirrorUCDFiles } from "./mirror";
-
 export {
   createLocalUCDStore,
   createRemoteUCDStore,
   createUCDStore,
+} from "./factory";
+
+export {
   UCDStore,
 } from "./store";
+
 export type {
-  AnalyzeResult,
-  CleanResult,
   LocalUCDStoreOptions,
   RemoteUCDStoreOptions,
+  StoreMode,
   UCDStoreOptions,
-} from "./store";
-export * from "./ucd-files";
+} from "./types";
