+++ packages/ucd-store/test/remote/error-handling.test.ts
@@ -1,9 +1,16 @@
 import type { FileSystemBridge } from "@ucdjs/utils/fs-bridge";
 import { HttpResponse, mockFetch, mockResponses } from "#msw-utils";
 import { UCDJS_API_BASE_URL } from "@ucdjs/env";
-import { flattenFilePaths } from "@ucdjs/ucd-store";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { createRemoteUCDStore } from "../../src/store";
+import {
+  afterEach,
+  beforeEach,
+  describe,
+  expect,
+  it,
+  vi,
+} from "vitest";
+import { createRemoteUCDStore } from "../../src/factory";
+import { flattenFilePaths } from "../../src/helpers";
 
 describe("remote ucd store - error handling", () => {
   let mockFs: FileSystemBridge;
