--- packages/ucd-store/test/store.test.ts
@@ -1,490 +1,0 @@
-import type { FileSystemBridge } from "@ucdjs/utils/fs-bridge";
-import { UCDJS_API_BASE_URL } from "@ucdjs/env";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { createLocalUCDStore, createUCDStore } from "../src/store";
-
-// eslint-disable-next-line test/prefer-lowercase-title
-describe("UCD Store - Common", () => {
-  let _mockFs: FileSystemBridge;
-
-  beforeEach(() => {
-    _mockFs = {
-      exists: vi.fn(),
-      read: vi.fn(),
-      write: vi.fn(),
-      listdir: vi.fn(),
-      mkdir: vi.fn(),
-      stat: vi.fn(),
-      rm: vi.fn(),
-    };
-  });
-
-  afterEach(() => {
-    vi.clearAllMocks();
-  });
-
-  describe("constructor", () => {
-    it("should create instance with default options", async () => {
-      const store = await createLocalUCDStore({
-        fs: _mockFs,
-      });
-
-      expect(store.mode).toBe("local");
-      expect(store.baseUrl).toBe(UCDJS_API_BASE_URL);
-    });
-
-    it("should create instance with custom options", async () => {
-      const customBaseUrl = "https://custom-base-url.com";
-      const store = await createLocalUCDStore({
-        fs: _mockFs,
-        baseUrl: customBaseUrl,
-      });
-
-      expect(store.baseUrl).toBe(customBaseUrl);
-    });
-
-    it("should set mode correctly", async () => {
-      const store = await createUCDStore({
-        fs: _mockFs,
-        mode: "local",
-      });
-
-      expect(store.mode).toBe("local");
-    });
-
-    it("should apply global filters", async () => {
-      const store = await createLocalUCDStore({
-        fs: _mockFs,
-        globalFilters: ["**Shaping.txt"],
-      });
-
-      expect(store.filter).toBeDefined();
-
-      expect(store.filter("ArabicShaping.txt")).toBe(true);
-
-      expect(store.filter("Other.txt")).toBe(false);
-    });
-  });
-
-  describe("version Management", () => {
-    it("should return loaded versions", () => {
-      // Test versions getter returns correct array
-    });
-
-    it("should check version existence correctly", () => {
-      // Test hasVersion method with existing and non-existing versions
-    });
-
-    it("should handle version list immutability", () => {
-      // Test that versions array cannot be modified externally
-    });
-  });
-
-  describe("filter Processing", () => {
-    it("should apply path filters correctly", async () => {
-      // Test that globalFilters work as expected
-    });
-
-    it("should handle empty filter patterns", async () => {
-      // Test behavior with no filters
-    });
-
-    it("should filter nested file structures", async () => {
-      // Test filtering of directory trees
-    });
-  });
-});
-
-// eslint-disable-next-line test/prefer-lowercase-title
-describe("UCD Store - Remote Mode", () => {
-  let _mockFs: FileSystemBridge;
-
-  beforeEach(() => {
-    // Setup mock filesystem bridge for remote mode
-    _mockFs = {
-      exists: vi.fn(),
-      read: vi.fn(),
-      write: vi.fn(),
-      listdir: vi.fn(),
-      mkdir: vi.fn(),
-      stat: vi.fn(),
-      rm: vi.fn(),
-    };
-  });
-
-  afterEach(() => {
-    vi.clearAllMocks();
-  });
-
-  describe("initialization", () => {
-    it("should initialize remote store successfully", async () => {
-      // Test that remote store loads UNICODE_VERSION_METADATA versions
-    });
-
-    it("should handle remote initialization errors", async () => {
-      // Test error handling during remote initialization
-    });
-  });
-
-  describe("file Operations", () => {
-    describe("getFileTree", () => {
-      it("should fetch file tree via API", async () => {
-        // Test API call and response processing for remote mode
-      });
-
-      it("should apply filters to remote file structure", async () => {
-        // Test that globalFilters are applied to file tree from API
-      });
-
-      it("should handle API errors", async () => {
-        // Test retry logic and error handling for API failures
-      });
-
-      it("should throw error for non-existent version", async () => {
-        // Test error when version is not in remote store
-      });
-    });
-
-    describe("getFile", () => {
-      it("should fetch file content via HTTP", async () => {
-        // Test file content retrieval via HTTP filesystem
-      });
-
-      it("should throw error for filtered file paths", async () => {
-        // Test that filtered files throw appropriate error in remote mode
-      });
-
-      it("should throw error for non-existent version", async () => {
-        // Test version validation in remote mode
-      });
-
-      it("should handle network errors", async () => {
-        // Test network error handling for file retrieval
-      });
-    });
-
-    describe("getFilePaths", () => {
-      it("should return flattened file paths from API", async () => {
-        // Test that remote file structure is properly flattened
-      });
-
-      it("should handle empty remote file trees", async () => {
-        // Test behavior with no files from API
-      });
-    });
-
-    describe("getAllFiles", () => {
-      it("should return files from all remote versions", async () => {
-        // Test aggregation across multiple versions from API
-      });
-
-      it("should prefix files with version paths", async () => {
-        // Test that returned remote paths include version prefix
-      });
-    });
-  });
-
-  describe("error Handling", () => {
-    it("should handle network connectivity issues", async () => {
-      // Test API connectivity problems
-    });
-
-    it("should handle API rate limiting", async () => {
-      // Test rate limit responses
-    });
-
-    it("should handle malformed API responses", async () => {
-      // Test invalid JSON or unexpected response format
-    });
-  });
-
-  describe("store Analysis", () => {
-    describe("analyze", () => {
-      it("should return correct analysis data for remote store", async () => {
-        // Test totalFiles count and versions list from remote
-      });
-
-      it("should handle empty remote stores", async () => {
-        // Test analysis of remote store with no files
-      });
-    });
-  });
-});
-
-// eslint-disable-next-line test/prefer-lowercase-title
-describe("UCD Store - Local Mode", () => {
-  let _mockFs: FileSystemBridge;
-
-  beforeEach(() => {
-    // Setup mock filesystem bridge for local mode
-    _mockFs = {
-      exists: vi.fn(),
-      read: vi.fn(),
-      write: vi.fn(),
-      listdir: vi.fn(),
-      mkdir: vi.fn(),
-      stat: vi.fn(),
-      rm: vi.fn(),
-    };
-  });
-
-  afterEach(() => {
-    vi.clearAllMocks();
-  });
-
-  describe("initialization", () => {
-    it("should initialize existing local store", async () => {
-      // Test loading from existing .ucd-store.json manifest
-    });
-
-    it("should create new local store with provided versions", async () => {
-      // Test creating new local store when directory doesn't exist
-    });
-
-    it("should throw error when no versions provided for new store", async () => {
-      // Test error when initializing new local store without versions
-    });
-
-    it("should throw error when base path is missing", async () => {
-      // Test error handling for missing basePath in local mode
-    });
-
-    it("should handle corrupted manifest file", async () => {
-      // Test error handling when .ucd-store.json is invalid JSON
-    });
-  });
-
-  describe("file Operations", () => {
-    describe("getFileTree", () => {
-      it("should read local directory structure", async () => {
-        // Test filesystem operations for local mode
-      });
-
-      it("should apply filters to local file structure", async () => {
-        // Test that globalFilters are applied to local directory tree
-      });
-
-      it("should throw error for non-existent version directory", async () => {
-        // Test error when version directory doesn't exist locally
-      });
-
-      it("should handle filesystem permission errors", async () => {
-        // Test handling of read permission issues
-      });
-    });
-
-    describe("getFile", () => {
-      it("should read file content from local filesystem", async () => {
-        // Test file reading from local filesystem
-      });
-
-      it("should throw error for filtered file paths", async () => {
-        // Test that filtered files throw appropriate error in local mode
-      });
-
-      it("should throw error for non-existent version", async () => {
-        // Test version validation in local mode
-      });
-
-      it("should handle file read errors", async () => {
-        // Test filesystem error handling for missing or corrupt files
-      });
-
-      it("should handle special characters in file paths", async () => {
-        // Test Unicode characters, spaces, etc. in local paths
-      });
-    });
-
-    describe("getFilePaths", () => {
-      it("should return flattened local file paths", async () => {
-        // Test that local file structure is properly flattened
-      });
-
-      it("should handle empty local directories", async () => {
-        // Test behavior with no local files
-      });
-    });
-
-    describe("getAllFiles", () => {
-      it("should return files from all local versions", async () => {
-        // Test aggregation across multiple local version directories
-      });
-
-      it("should prefix files with version paths", async () => {
-        // Test that returned local paths include version prefix
-      });
-    });
-  });
-
-  describe("manifest Operations", () => {
-    it("should create valid store manifest", async () => {
-      // Test .ucd-store.json creation with correct structure
-    });
-
-    it("should load manifest correctly", async () => {
-      // Test parsing of existing manifest file
-    });
-
-    it("should handle manifest with invalid structure", async () => {
-      // Test error handling for malformed manifests
-    });
-
-    it("should update manifest when versions change", async () => {
-      // Test manifest updates during store operations
-    });
-  });
-
-  describe("error Handling", () => {
-    it("should handle filesystem bridge errors", async () => {
-      // Test various local filesystem operation failures
-    });
-
-    it("should handle disk space issues", async () => {
-      // Test behavior when filesystem operations fail due to space
-    });
-
-    it("should handle concurrent access to manifest", async () => {
-      // Test thread safety for manifest operations
-    });
-  });
-
-  describe("store Analysis", () => {
-    describe("analyze", () => {
-      it("should return correct analysis data for local store", async () => {
-        // Test totalFiles count and versions list from local filesystem
-      });
-
-      it("should handle empty local stores", async () => {
-        // Test analysis of local store with no files
-      });
-    });
-
-    describe("clean", () => {
-      it("should throw not implemented error", async () => {
-        // Test that clean method throws expected error for local store
-      });
-    });
-  });
-});
-
-describe("createUCDStore Factory Function", () => {
-  describe("remote mode factory", () => {
-    it("should create and initialize remote store successfully", async () => {
-      // Test factory function creates initialized remote store
-    });
-
-    it("should pass remote options correctly to constructor", async () => {
-      // Test that remote-specific options are forwarded properly
-    });
-
-    it("should handle remote initialization failures", async () => {
-      // Test error propagation from remote initialization
-    });
-  });
-
-  describe("local mode factory", () => {
-    it("should create and initialize local store successfully", async () => {
-      // Test factory function creates initialized local store
-    });
-
-    it("should pass local options correctly to constructor", async () => {
-      // Test that local-specific options are forwarded properly
-    });
-
-    it("should handle local initialization failures", async () => {
-      // Test error propagation from local initialization
-    });
-  });
-});
-
-describe("integration Tests", () => {
-  describe("remote mode integration", () => {
-    it("should handle complete remote workflow", async () => {
-      // Test end-to-end remote store operations
-    });
-
-    it("should handle remote store with filtering", async () => {
-      // Test remote store operations with global filters applied
-    });
-
-    it("should handle remote store error recovery", async () => {
-      // Test resilience to network issues and API failures
-    });
-  });
-
-  describe("local mode integration", () => {
-    it("should handle complete local workflow", async () => {
-      // Test end-to-end local store operations
-    });
-
-    it("should handle local store with filtering", async () => {
-      // Test local store operations with global filters applied
-    });
-
-    it("should handle local store migration", async () => {
-      // Test upgrading or migrating local store structure
-    });
-  });
-
-  describe("cross-mode compatibility", () => {
-    it("should handle data compatibility between modes", async () => {
-      // Test that data structures are compatible between remote and local
-    });
-
-    it("should handle switching between modes", async () => {
-      // Test scenario where store mode changes
-    });
-  });
-});
-
-describe("edge Cases and Performance", () => {
-  describe("performance tests", () => {
-    it("should handle large remote file trees efficiently", async () => {
-      // Test performance with many files from API
-    });
-
-    it("should handle large local file trees efficiently", async () => {
-      // Test performance with many local files
-    });
-
-    it("should handle concurrent remote operations", async () => {
-      // Test concurrent API calls and caching
-    });
-
-    it("should handle concurrent local operations", async () => {
-      // Test concurrent filesystem operations
-    });
-  });
-
-  describe("edge cases", () => {
-    it("should handle special characters in remote file paths", async () => {
-      // Test Unicode characters, spaces, etc. in remote paths
-    });
-
-    it("should handle special characters in local file paths", async () => {
-      // Test Unicode characters, spaces, etc. in local paths
-    });
-
-    it("should handle very deep directory structures", async () => {
-      // Test deeply nested directories in both modes
-    });
-
-    it("should handle empty stores gracefully", async () => {
-      // Test behavior with no versions or files in both modes
-    });
-  });
-
-  describe("resource management", () => {
-    it("should handle remote API rate limiting", async () => {
-      // Test behavior under API rate limits
-    });
-
-    it("should handle local disk space issues", async () => {
-      // Test behavior when local filesystem runs out of space
-    });
-
-    it("should clean up resources properly", async () => {
-      // Test resource cleanup in both modes
-    });
-  });
-});
