+++ packages/ucd-store/tsdown.config.ts
@@ -3,6 +3,5 @@ import { createTsdownConfig } from "@ucdjs/tsdown-config";
 export default createTsdownConfig({
   entry: [
     "./src/index.ts",
-    "./src/ucd-files.ts",
   ],
 });
